// src/UrlTracker.Backoffice.UI/api-client/types.gen.ts
export type CreateRedirectRequest = {
  source: StrategyViewModel;
  target: StrategyViewModel;
  permanent: boolean;
  retainQuery: boolean;
  force: boolean;
  advanced: boolean;
  key?: string | null;
  solvedRecommendation?: number | null;
};
// ...
export type RedirectRequest = {
  source: StrategyViewModel;
  target: StrategyViewModel;
  permanent: boolean;
  retainQuery: boolean;
  force: boolean;
  advanced: boolean;
};
// ...
export type StrategyViewModel = {
  strategy: string;
  value: string;
};